refactor: Extract optimistic drag/drop helper
