+++ force-app/main/default/lwc/lresKanbanExplorer/boardInteractions.js
@@ -1,6 +1,7 @@
 import { NavigationMixin } from "lightning/navigation";
 import { updateRecord } from "lightning/uiRecordApi";
 import { buildFilterDefinitions as buildFilterDefinitionsUtil } from "./filterUtils";
+import { buildOptimisticColumnsForDrop } from "./dragDropUtils";
 import { sanitizeFieldOutput } from "c/lresOutputUtils";
 
 export function buildFilterDefinitions(component, records) {
@@ -730,76 +731,6 @@ function scheduleSearchRebuild(component) {
   component.rebuildColumnsWithPicklist();
 }
 
-function buildOptimisticColumnsForDrop(
-  component,
-  { recordId, sourceColumnKey, targetColumnKey }
-) {
-  const columns = Array.isArray(component.columns) ? component.columns : [];
-  if (!columns.length) {
-    return null;
-  }
-
-  const targetColumn = component.findColumnByKey(targetColumnKey);
-  if (!targetColumn) {
-    return null;
-  }
-
-  let sourceColumn = sourceColumnKey
-    ? component.findColumnByKey(sourceColumnKey)
-    : null;
-  let sourceCard = null;
-  if (sourceColumn && Array.isArray(sourceColumn.records)) {
-    sourceCard = sourceColumn.records.find((card) => card.id === recordId);
-  }
-
-  if (!sourceCard) {
-    sourceColumn =
-      columns.find(
-        (column) =>
-          Array.isArray(column.records) &&
-          column.records.some((card) => card.id === recordId)
-      ) || null;
-    sourceCard = sourceColumn?.records?.find((card) => card.id === recordId);
-  }
-
-  if (!sourceColumn || !sourceCard) {
-    return null;
-  }
-
-  if (sourceColumn.key === targetColumn.key) {
-    return null;
-  }
-
-  const savingCard = { ...sourceCard, isSaving: true };
-  const nextColumns = columns.map((column) => {
-    if (column.key === sourceColumn.key) {
-      const records = Array.isArray(column.records) ? column.records : [];
-      const nextRecords = records.filter((card) => card.id !== recordId);
-      return {
-        ...column,
-        records: nextRecords
-      };
-    }
-    if (column.key === targetColumn.key) {
-      const records = Array.isArray(column.records) ? column.records : [];
-      const nextRecords = [
-        ...records.filter((card) => card.id !== recordId),
-        savingCard
-      ];
-      return {
-        ...column,
-        records: nextRecords
-      };
-    }
-    return column;
-  });
-
-  return {
-    previousColumns: columns,
-    nextColumns
-  };
-}
-
 export async function updateRecordGrouping(
   component,
   { recordId, sourceColumnKey, targetColumnKey, blankKey }
@@ -826,10 +757,11 @@ export async function updateRecordGrouping(
   fields[groupingField] =
     newValue === null || newValue === undefined ? null : String(newValue);
 
-  const optimisticColumns = buildOptimisticColumnsForDrop(component, {
+  const optimisticColumns = buildOptimisticColumnsForDrop(component.columns, {
     recordId,
     sourceColumnKey,
-    targetColumnKey
+    targetColumnKey,
+    findColumnByKey: (key) => component.findColumnByKey(key)
   });
   if (optimisticColumns) {
     component.columns = optimisticColumns.nextColumns;
