+++ force-app/main/default/lwc/lresKanbanExplorer/lresKanbanExplorer.html
@@ -1,6 +1,6 @@
 <template>
   <lightning-card title={resolvedBoardTitle} icon-name="standard:kanban">
-    <div class="kanban-version">drift-silver-ember</div>
+    <div class="kanban-version">signal-cloud-ember</div>
     <template if:true={actionsAvailable}>
       <c-lres-kanban-board-actions
         slot="actions"
